consolidate columns module
