--- db/columns/defaults.py
@@ -1,4 +1,0 @@
-NAME = "name"
-DESCRIPTION = "description"
-NULLABLE = "nullable"
-DEFAULT = "default"
